+++ tests/Reader/Iterable/FilterHandler/GreaterThanOrEqualTest.php
@@ -2,22 +2,22 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Data\Tests\Reader\IterableHandler;
+namespace Yiisoft\Data\Tests\Reader\Iterable\FilterHandler;
 
 use DateTimeImmutable;
 use PHPUnit\Framework\Attributes\DataProvider;
-use Yiisoft\Data\Reader\Filter\GreaterThan;
-use Yiisoft\Data\Reader\Iterable\FilterHandler\GreaterThanHandler;
+use Yiisoft\Data\Reader\Filter\GreaterThanOrEqual;
+use Yiisoft\Data\Reader\Iterable\FilterHandler\GreaterThanOrEqualHandler;
 use Yiisoft\Data\Tests\TestCase;
 
-final class GreaterThanTest extends TestCase
+final class GreaterThanOrEqualTest extends TestCase
 {
     public static function matchScalarDataProvider(): array
     {
         return [
             [true, 44],
-            [true, '44'],
-            [false, 45],
+            [true, 45],
+            [true, '45'],
             [false, 46],
         ];
     }
@@ -25,21 +25,21 @@ final class GreaterThanTest extends TestCase
     #[DataProvider('matchScalarDataProvider')]
     public function testMatchScalar(bool $expected, mixed $value): void
     {
-        $handler = new GreaterThanHandler();
+        $processor = new GreaterThanOrEqualHandler();
 
         $item = [
             'id' => 1,
             'value' => 45,
         ];
 
-        $this->assertSame($expected, $handler->match($item, new GreaterThan('value', $value), []));
+        $this->assertSame($expected, $processor->match($item, new GreaterThanOrEqual('value', $value), []));
     }
 
     public static function matchDateTimeInterfaceDataProvider(): array
     {
         return [
             [true, new DateTimeImmutable('2022-02-22 16:00:44')],
-            [false, new DateTimeImmutable('2022-02-22 16:00:45')],
+            [true, new DateTimeImmutable('2022-02-22 16:00:45')],
             [false, new DateTimeImmutable('2022-02-22 16:00:46')],
         ];
     }
@@ -47,13 +47,13 @@ final class GreaterThanTest extends TestCase
     #[DataProvider('matchDateTimeInterfaceDataProvider')]
     public function testMatchDateTimeInterface(bool $expected, DateTimeImmutable $value): void
     {
-        $handler = new GreaterThanHandler();
+        $processor = new GreaterThanOrEqualHandler();
 
         $item = [
             'id' => 1,
             'value' => new DateTimeImmutable('2022-02-22 16:00:45'),
         ];
 
-        $this->assertSame($expected, $handler->match($item, new GreaterThan('value', $value), []));
+        $this->assertSame($expected, $processor->match($item, new GreaterThanOrEqual('value', $value), []));
     }
 }
