+++ tests/Reader/Iterable/FilterHandler/GreaterThanTest.php
@@ -2,22 +2,22 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Data\Tests\Reader\IterableHandler;
+namespace Yiisoft\Data\Tests\Reader\Iterable\FilterHandler;
 
 use DateTimeImmutable;
 use PHPUnit\Framework\Attributes\DataProvider;
-use Yiisoft\Data\Reader\Filter\GreaterThanOrEqual;
-use Yiisoft\Data\Reader\Iterable\FilterHandler\GreaterThanOrEqualHandler;
+use Yiisoft\Data\Reader\Filter\GreaterThan;
+use Yiisoft\Data\Reader\Iterable\FilterHandler\GreaterThanHandler;
 use Yiisoft\Data\Tests\TestCase;
 
-final class GreaterThanOrEqualTest extends TestCase
+final class GreaterThanTest extends TestCase
 {
     public static function matchScalarDataProvider(): array
     {
         return [
             [true, 44],
-            [true, 45],
-            [true, '45'],
+            [true, '44'],
+            [false, 45],
             [false, 46],
         ];
     }
@@ -25,21 +25,21 @@ final class GreaterThanOrEqualTest extends TestCase
     #[DataProvider('matchScalarDataProvider')]
     public function testMatchScalar(bool $expected, mixed $value): void
     {
-        $processor = new GreaterThanOrEqualHandler();
+        $handler = new GreaterThanHandler();
 
         $item = [
             'id' => 1,
             'value' => 45,
         ];
 
-        $this->assertSame($expected, $processor->match($item, new GreaterThanOrEqual('value', $value), []));
+        $this->assertSame($expected, $handler->match($item, new GreaterThan('value', $value), []));
     }
 
     public static function matchDateTimeInterfaceDataProvider(): array
     {
         return [
             [true, new DateTimeImmutable('2022-02-22 16:00:44')],
-            [true, new DateTimeImmutable('2022-02-22 16:00:45')],
+            [false, new DateTimeImmutable('2022-02-22 16:00:45')],
             [false, new DateTimeImmutable('2022-02-22 16:00:46')],
         ];
     }
@@ -47,13 +47,13 @@ final class GreaterThanOrEqualTest extends TestCase
     #[DataProvider('matchDateTimeInterfaceDataProvider')]
     public function testMatchDateTimeInterface(bool $expected, DateTimeImmutable $value): void
     {
-        $processor = new GreaterThanOrEqualHandler();
+        $handler = new GreaterThanHandler();
 
         $item = [
             'id' => 1,
             'value' => new DateTimeImmutable('2022-02-22 16:00:45'),
         ];
 
-        $this->assertSame($expected, $processor->match($item, new GreaterThanOrEqual('value', $value), []));
+        $this->assertSame($expected, $handler->match($item, new GreaterThan('value', $value), []));
     }
 }
